Revert "..."
This reverts commit 95017a0df036894de496d1bcb94dbf033a6390dc.

+++ application/controllers/Datasiswa.php
@@ -16,7 +16,7 @@ class Multiple extends CI_Controller
         $this->data['users'] = $this->db->get('users')->result();
         $this->load->view('header_view', $this->data);
         $this->load->view('sidebar_view', $this->data);
-        $this->load->view('multiple/index', $this->data);
+        $this->load->view('siswa/index', $this->data);
         $this->load->view('footer_view', $this->data);
         // $this->load->view('multiple/index', $this->data);
     }
@@ -107,13 +107,18 @@ class Multiple extends CI_Controller
             redirect('multiple/index');
         }
     }
-    public function delete()
+    public function delete($id)
     {
-
-        $id = htmlspecialchars($this->input->post('id'));
-        $this->db->where('id', $id);
-        $this->db->delete('users');
-        redirect('Multiple');
+        if ($id == NULL) {
+            $this->db->get_where('users', ['id' => $id])->result_array();
+        } else {
+            return $this->db->get_where('users', ['id' => $id])->row_array();
+            $id = htmlspecialchars($this->input->post('id'));
+            $this->db->where('id', $id);
+            $this->db->delete('users');
+            $this->sesion->set_flasdata('users', 'Telah Dihapus');
+            redirect('Multiple');
+        }
     }
     public function edit($id)
     {
@@ -128,66 +133,4 @@ class Multiple extends CI_Controller
         $this->load->view('siswa/edit', $this->data);
         $this->load->view('footer_view', $this->data);
         // $this->load->view('multiple/edit', $data);
-    }
-
-    public function approval($id)
-    {
-        $a = $this->input->post('kode');
-        $b = $this->input->post('status');
-        // var_dump($this->session->userdata());
-        // die;
-        // var_dump($a);
-        // var_dump($b);
-        // die;
-        $i = 1;
-        $idus = $id;
-        $kd = $a[$i];
-
-        date_default_timezone_set('Asia/Jakarta');
-        $data = [
-            'kode' => $this->input->post('kode'),
-            'status' => $this->input->post('status'),
-            'created_date' => date('Y-m-d H:i:s'),
-        ];
-        // $this->db->where('id_user', $idus);
-        $this->db->where('kode', $a);
-        $this->db->update('kode_area', $data);
-        $where = [
-            'id_user' => $id,
-            'kode' => $a[$i]
-        ];
-        // var_dump($where);
-        // die;
-        // $data = ['status' => $b];
-        // var_dump($data);
-        // $this->db->last_query();
-        // die;
-
-        foreach ($b as $status) {
-            if (!empty($status)) {
-                $idus = $id;
-                $kd = $a[$i];
-
-                // $where = [
-                //     'id_user' => $id,
-                //     'kode' => $a[$i]
-                // ];
-                // var_dump($where);
-                // die;
-                $data = ['status' => $status];
-                var_dump($data);
-                $this->db->where('id_user', $idus);
-                $this->db->where('id_user', $kd);
-                $this->db->update('kode_area', $data);
-                // $this->db->last_query();
-                // die;
-                // $this->Mcrud->update('kode_area', $data, $where);
-                // $i++;
-            }
-        }
-
-        $this->Mcrud->update();
-        $this->session->set_flashdata('pesan', 'Data area user berhasil diverifikasi');
-        redirect('index.php/multiple');
-    }
-}
+    }ssss
